# backend/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  github_com_blagovibe_TorrSyncPlayer_backend_internal_models.CreateRoomRequest:
    properties:
      name:
        type: string
      password:
        type: string
    type: object
  github_com_blagovibe_TorrSyncPlayer_backend_internal_models.ErrorResponse:
    properties:
      error:
        type: string
    type: object
...
  github_com_blagovibe_TorrSyncPlayer_backend_internal_models.JoinRoomRequest:
    properties:
      password:
        type: string
      roomId:
        type: string
    type: object
...
  github_com_blagovibe_TorrSyncPlayer_backend_internal_models.RoomInfo:
    properties:
      hostId:
        description: |-
          HostID is the host's peer (session) identifier, NOT the host's user ID.
          The host's user ID is tracked separately server-side (HostUserID in the
          p2p service room struct) and is the authoritative key for host checks.
        type: string
      id:
        type: string
      name:
        type: string
      peerCount:
        type: integer
    type: object
  github_com_blagovibe_TorrSyncPlayer_backend_internal_models.SeekRequest:
    properties:
      position:
        type: number
    type: object
...
  github_com_blagovibe_TorrSyncPlayer_backend_internal_models.SignalRequest:
    properties:
      roomId:
        type: string
      signal:
        items:
          type: integer
        type: array
    type: object
...
  github_com_blagovibe_TorrSyncPlayer_backend_internal_models.SuccessResponse:
    properties:
      message:
        type: string
    type: object
  github_com_blagovibe_TorrSyncPlayer_backend_internal_models.SyncStatus:
    properties:
      duration:
        type: number
      isPlaying:
        type: boolean
      position:
        type: number
      timestamp:
        type: integer
    type: object
...
host: localhost:8889
info:
  contact:
    email: support@torrsyncplayer.local
    name: TorrSyncPlayer Support
    url: https://github.com/blagovibe/TorrSyncPlayer/issues
  description: HTTP API for TorrSyncPlayer — torrent player with P2P playback synchronization.
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: https://github.com/blagovibe/TorrSyncPlayer
  title: TorrSyncPlayer API
  version: "1.0"
paths:
...
  /api/v1/rooms:
    post:
      consumes:
      - application/json
      description: Creates a new P2P room for playback synchronization
      parameters:
      - description: Room data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_blagovibe_TorrSyncPlayer_backend_internal_models.CreateRoomRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/github_com_blagovibe_TorrSyncPlayer_backend_internal_models.RoomInfo'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_blagovibe_TorrSyncPlayer_backend_internal_models.ErrorResponse'
      summary: Create room
      tags:
      - rooms
  /api/v1/rooms/{roomID}/events:
    get:
      description: Subscribe to P2P room events in real-time via Server-Sent Events
      parameters:
      - description: Room ID
        in: path
        name: roomID
        required: true
        type: string
      produces:
      - text/event-stream
      responses:
        "200":
          description: OK
          schema:
            type: string
      summary: Room events (SSE)
      tags:
      - rooms
  /api/v1/rooms/join:
    post:
      consumes:
      - application/json
      description: Joins a user to an existing P2P room
      parameters:
      - description: Login data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_blagovibe_TorrSyncPlayer_backend_internal_models.JoinRoomRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_blagovibe_TorrSyncPlayer_backend_internal_models.SuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_blagovibe_TorrSyncPlayer_backend_internal_models.ErrorResponse'
      summary: Join room
      tags:
      - rooms
  /api/v1/rooms/leave:
    post:
      description: Leaves the current P2P room
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_blagovibe_TorrSyncPlayer_backend_internal_models.SuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_blagovibe_TorrSyncPlayer_backend_internal_models.ErrorResponse'
      summary: Leave room
      tags:
      - rooms
  /api/v1/rooms/signal:
    post:
      consumes:
      - application/json
      description: Relays a sync signal (arbitrary JSON payload) to all peers in the
        room via the server-brokered SSE stream
      parameters:
      - description: Sync signal
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_blagovibe_TorrSyncPlayer_backend_internal_models.SignalRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_blagovibe_TorrSyncPlayer_backend_internal_models.SuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_blagovibe_TorrSyncPlayer_backend_internal_models.ErrorResponse'
      summary: Send room sync signal
      tags:
      - rooms
...
  /api/v1/sync/seek:
    post:
      consumes:
      - application/json
      description: Performs synchronized seek to the specified position
      parameters:
      - description: Seek position
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_blagovibe_TorrSyncPlayer_backend_internal_models.SeekRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_blagovibe_TorrSyncPlayer_backend_internal_models.SyncStatus'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_blagovibe_TorrSyncPlayer_backend_internal_models.ErrorResponse'
      summary: Seek
      tags:
      - sync
...
securityDefinitions:
  BearerAuth:
    description: 'Enter JWT token in the format: Bearer <token>'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"